fix bug in parsing layer; also, add tests
