+++ zkstats/onnx2circom/__init__.py
@@ -33,11 +33,12 @@ def keras_to_circom(keras_path: Path, generated_circom_path: Path):
     #     # TODO: should we skip them?
     #     skips=['util.circom', 'circomlib-matrix', 'circomlib', 'crypto'],
     # )
-    circom.file_parse(MPC_CIRCOM_PATH)
+    templates = circom.file_parse(MPC_CIRCOM_PATH)
     # transpiler.transpile(args['<model.h5>'], args['--output'], args['--raw'], args['--decimals'])
     # keras2circom_output_dir = Path(tempfile.mkdtemp())
     keras2circom_output_dir = generated_circom_path.parent
     transpiler.transpile(
+        templates,
         str(keras_path),
         str(keras2circom_output_dir),
     )
