+++ zkstats/onnx2circom/keras2circom/main.py
@@ -18,8 +18,8 @@ from keras2circom import circom, transpiler
 def main():
     """ Main entry point of the app """
     args = docopt(__doc__)
-    circom.dir_parse('node_modules/circomlib-ml/circuits/', skips=['util.circom', 'circomlib-matrix', 'circomlib', 'crypto'])
-    transpiler.transpile(args['<model.h5>'], args['--output'], args['--raw'], args['--decimals'])
+    templates = circom.dir_parse('node_modules/circomlib-ml/circuits/', skips=['util.circom', 'circomlib-matrix', 'circomlib', 'crypto'])
+    transpiler.transpile(templates, args['<model.h5>'], args['--output'], args['--raw'], args['--decimals'])
 
 if __name__ == "__main__":
     """ This is executed when run from the command line """
